+++ client/src/extension.ts
@@ -41,22 +41,54 @@ export function activate(context: ExtensionContext) {
 
 	let yseopCliPath : string = vscode.workspace.getConfiguration('yseopml').get('pathToYseopCli');
 
-	const batch = vscode.commands.registerCommand('yseopml.batch', () =>
-    {
+	const batchCmd = vscode.commands.registerCommand('yseopml.batch', () => {
 		ExecYseopCliCommand(yseopCliPath, "batch");
-    });
+	});
 	
-	const compile = vscode.commands.registerCommand('yseopml.compile', () =>
-    {
+	const compileCmd = vscode.commands.registerCommand('yseopml.compile', () => {
 		ExecYseopCliCommand(yseopCliPath, "compile");
-    });
+	});
+	
+	const testCmd = vscode.commands.registerCommand('yseopml.test', () => {
+		ExecYseopCliCommand(yseopCliPath, "test");
+	});
+	
+	const cleanCmd = vscode.commands.registerCommand('yseopml.clean', () => {
+		ExecYseopCliCommand(yseopCliPath, "clean");
+	});
+	
+	const cleanallCmd = vscode.commands.registerCommand('yseopml.cleanall', () => {
+		ExecYseopCliCommand(yseopCliPath, "clean", "--all");
+	});
+	
+	const packageCmd = vscode.commands.registerCommand('yseopml.package', () => {
+		ExecYseopCliCommand(yseopCliPath, "package");
+	});
+	
+	const infoCmd = vscode.commands.registerCommand('yseopml.info', () => {
+		ExecYseopCliCommand(yseopCliPath, "info");
+	});
 
 	// Create the language client and start the client.
 	console.log("Yseop.vscode-yseopml − Starting Language Client");
-	let disposable = new LanguageClient('yseopml', 'Yseop Markup Language language server', serverOptions, clientOptions, true).start();
+	let disposable = new LanguageClient(
+			'yseopml', 'Yseop Markup Language language server',
+			serverOptions, clientOptions, true
+	).start();
+
 	// Push the disposable to the context's subscriptions so that the 
-	// client can be deactivated on extension deactivation
-	context.subscriptions.push(disposable, batch, compile);
+	// client can be deactivated on extension deactivation.
+	// Also register the custom commands.
+	context.subscriptions.push(
+		disposable,
+		batchCmd,
+		compileCmd,
+		testCmd,
+		cleanCmd,
+		cleanallCmd,
+		packageCmd,
+		infoCmd
+	);
 }
 
 /**
@@ -64,7 +96,7 @@ export function activate(context: ExtensionContext) {
  * @param yseopCliPath The absolute path to Yseop CLI executable.
  * @param commandName The Yseop CLI subcommand to use, like ”batch” or ”test”.
  */
-export async function ExecYseopCliCommand(yseopCliPath: string, commandName: string) {
+export async function ExecYseopCliCommand(yseopCliPath: string, ...words: string[]) {
 	const editor = vscode.window.activeTextEditor;
 	
 	if(isNullOrUndefined(editor)) {
@@ -89,7 +121,11 @@ export async function ExecYseopCliCommand(yseopCliPath: string, commandName: str
 	const workspaceFolders = vscode.workspace.workspaceFolders;
 	const kbDirectory = workspaceFolders[0].uri.fsPath;
 	
-	const commandLine = `"${yseopCliPath}" ${commandName} "${kbDirectory}"`;
+	var commandLine = `"${yseopCliPath}"`;
+	words.forEach((oneWord) => {
+		commandLine += ` "${oneWord}"`;
+	});
+	commandLine += ` "${kbDirectory}"`;
 
 	const command = exec(commandLine);